LayoutTests/http/tests/adClickAttribution/attribution-conversion-through-image-redirect-without-priority-expected.txt

Tests triggering of ad click attribution conversions without priority.


WebCore::AdClickAttribution 1
Source: 127.0.0.1
Destination: localhost
Campaign ID: 3
Conversion data: 12
Conversion priority: 0
